feat: increasing the size of vector and matrix coordinates

+++ include/matriz.hpp
@@ -5,14 +5,14 @@
 
 class Matriz {
 	private:
-		unsigned int a00, a01, a10, a11;
+		unsigned long int a00, a01, a10, a11;
 	public:
-		Matriz(unsigned int newA00, unsigned int newA01, unsigned int newA10, unsigned int newA11);
-		unsigned int getA00();
-		unsigned int getA01();
-		unsigned int getA10();
-		unsigned int getA11();
-		void atualizaValores(unsigned int newA00, unsigned int newA01, unsigned int newA10, unsigned int newA11);
+		Matriz(unsigned long int newA00, unsigned long int newA01, unsigned long int newA10, unsigned long int newA11);
+		unsigned long int getA00();
+		unsigned long int getA01();
+		unsigned long int getA10();
+		unsigned long int getA11();
+		void atualizaValores(unsigned long int newA00, unsigned long int newA01, unsigned long int newA10, unsigned long int newA11);
 		void transformaVetor(Vetor* vetor);
 		static void calculaProduto(Matriz* a, Matriz* b, Matriz* resultado);
 };
