+++ src/main.cpp
@@ -17,13 +17,13 @@ int main (int argc, char *argv[]) {
 			std::cin >> entrada;
 			unsigned int instante = std::stoi(entrada);
 			std::cin >> entrada;
-			unsigned int a00 = std::stoi(entrada);
+			unsigned long int a00 = std::stoi(entrada);
 			std::cin >> entrada;
-			unsigned int a01 = std::stoi(entrada);
+			unsigned long int a01 = std::stoi(entrada);
 			std::cin >> entrada;
-			unsigned int a10 = std::stoi(entrada);
+			unsigned long int a10 = std::stoi(entrada);
 			std::cin >> entrada;
-			unsigned int a11 = std::stoi(entrada);
+			unsigned long int a11 = std::stoi(entrada);
 			Matriz* transformacao = new Matriz(a00, a01, a10, a11);
 			jogo.atualiza(instante, transformacao);
 		} else if(entrada == "q") {
@@ -32,12 +32,12 @@ int main (int argc, char *argv[]) {
 			std::cin >> entrada;
 			unsigned int morte = std::stoi(entrada);
 			std::cin >> entrada;
-			unsigned int x = std::stoi(entrada);
+			unsigned long int x = std::stoi(entrada);
 			std::cin >> entrada;
-			unsigned int y = std::stoi(entrada);
+			unsigned long int y = std::stoi(entrada);
 			Vetor* vetor = new Vetor(x, y);
 			jogo.consulta(nascimento, morte, vetor);
-			std::cout << vetor->getX() % 100000000 << " " << vetor->getY() % 100000000 << "\n";
+			std::cout << vetor->getX() % 100000000L << " " << vetor->getY() % 100000000L << "\n";
 			delete vetor;
 		}
 	}
